<chore> 서버 분리에 따른 디렉토리 구조 정리
