+++ virspit-auth/src/main/java/com/virspit/virspitauth/controller/AdviceController.java
@@ -1,4 +1,4 @@
-package com.virspit.virspitauth.domain.member.controller;
+package com.virspit.virspitauth.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
