--- virspit-auth/src/main/java/com/virspit/virspitauth/domain/member/dto/request/MemberSignUpRequestDto.java
@@ -1,35 +1,0 @@
-package com.virspit.virspitauth.domain.member.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
-import com.virspit.virspitauth.domain.member.entity.Gender;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import org.apache.tomcat.jni.Local;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import javax.validation.constraints.NotNull;
-import java.time.LocalDate;
-import java.util.Date;
-
-@Getter
-@Setter
-@ToString
-public class MemberSignUpRequestDto {
-
-    private String username;
-
-    private String email;
-
-    private String password;
-
-    private Gender gender;
-
-    private LocalDate birthdayDate;
-
-}
-
